+++ pipeline/__init__.py
@@ -1,0 +1,1 @@
+"""Pipeline microservice package."""